Fix issues when staging binary files
- In Hasher, we were copying the binary file into bins/ directory all the time even when we just want to get the hash. So, we will copy it only if the files doesn't exist
- In FileOps, while returning the lines for the binary file, we were sending in the hash code and somehow the hash code is not coming out unique even when the file doesn't change. So, instead we will take 100 elements of array and use that as the line of that binary file
- A few more changes In Stage,  to make sure it works fine all the time

+++ src/main/scala/delorean/Hasher.scala
@@ -80,7 +80,7 @@ class Hasher {
 
         // When its a binary file, don't do all the usual line extractions and hashing.
         // Just put the file into BINARIES_FOLDER with the filehash as the name
-        if (isBinaryFile(filePath)) {
+        if (isBinaryFile(filePath) && Files.notExists(Paths.get(BINARIES_FOLDER + fileHash))) {
             copyFile(filePath, BINARIES_FOLDER + fileHash)
 
             // Once the file hash is computed, Add it to travelogue file
